renames/added the cpp files so they match the sections of swephprg.doc
