+++ R/RcppExports.R
@@ -1,6 +1,66 @@
 # Generated by using Rcpp::compileAttributes() -> do not edit by hand
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
+#' @title Expert interface: miscellanious functions
+#'
+#' @description For experts that are familiar with the C API of the underlying
+#'   \code{libswe}, an expert interface that mimicks that interface as closely
+#'   as possbile is provided.
+#' @param jd  Julian day number as numeric vector (day)
+#' @param t_acc tidal acceleration as double (arcsec/century^2)
+#' @param path  directory as string, where the ephemeris files are stored
+#' @param delta_t DeltaT (day)
+#' @return \code{swe_day_of_week} retruns the day of week as integer vector (0 Monday .. 6 Sunday)
+#' @return \code{swe_get_tid_acc} returns the tidal acceleration as double (arcsec/century^2)
+#' @return \code{swe_version} returns the Swiss Ephemeris version number as string
+#' @return \code{swe_deltat} returns the DeltaT as double (day)
+#' @name expert-interface
+NULL
+
+#' Get the Swiss Ephemeris version number
+#' @rdname expert-interface
+#' @export
+swe_version <- function() {
+    .Call(`_swephR_version`)
+}
+
+#' Set the directory for the sefstars.txt, swe_deltat.txt and jpl files
+#' @rdname expert-interface
+#' @export
+swe_set_ephe_path <- function(path) {
+    invisible(.Call(`_swephR_set_ephe_path`, path))
+}
+
+#' Close Swiss Ephemeris files
+#' @rdname expert-interface
+#' @export
+swe_close <- function() {
+    invisible(.Call(`_swephR_close`))
+}
+
+#' @title Expert interface: miscellanious functions
+#'
+#' @description For experts that are familiar with the C API of the underlying
+#'   \code{libswe}, an expert interface that mimicks that interface as closely
+#'   as possbile is provided.
+#' @param jd  Julian day number as numeric vector (day)
+#' @param t_acc tidal acceleration as double (arcsec/century^2)
+#' @param path  directory as string, where the ephemeris files are stored
+#' @param delta_t DeltaT (day)
+#' @return \code{swe_day_of_week} retruns the day of week as integer vector (0 Monday .. 6 Sunday)
+#' @return \code{swe_get_tid_acc} returns the tidal acceleration as double (arcsec/century^2)
+#' @return \code{swe_version} returns the Swiss Ephemeris version number as string
+#' @return \code{swe_deltat} returns the DeltaT as double (day)
+#' @name expert-interface
+NULL
+
+#' Compute day of week
+#' @rdname expert-interface
+#' @export
+swe_day_of_week <- function(jd) {
+    .Call(`_swephR_day_of_week`, jd)
+}
+
 calc_ut <- function(jd_ut, ipl, iflag) {
     .Call(`_swephR_calc_ut`, jd_ut, ipl, iflag)
 }
@@ -9,31 +69,21 @@ calc <- function(jd_et, ipl, iflag) {
     .Call(`_swephR_calc`, jd_et, ipl, iflag)
 }
 
-fixstar2_ut <- function(star, jd_ut, iflag) {
-    .Call(`_swephR_fixstar2_ut`, star, jd_ut, iflag)
+fixstar2_ut <- function(starname, jd_ut, iflag) {
+    .Call(`_swephR_fixstar2_ut`, starname, jd_ut, iflag)
 }
 
-fixstar2 <- function(star, jd_et, iflag) {
-    .Call(`_swephR_fixstar2`, star, jd_et, iflag)
+fixstar2 <- function(starname, jd_et, iflag) {
+    .Call(`_swephR_fixstar2`, starname, jd_et, iflag)
 }
 
 #' Calculation magnitude of star
 #' @return \code{swe_fixstar2_mag} returns a list with named entries \code{return} status flag as interger,
-#'         \code{star} updated star name as string, \code{mag} magnitude of star as double, and \code{serr} for error message as string.
+#'         \code{starname} updated star name as string, \code{mag} magnitude of star as double, and \code{serr} for error message as string.
 #' @rdname calculation
 #' @export
-swe_fixstar2_mag <- function(star) {
-    .Call(`_swephR_fixstar2_mag`, star)
-}
-
-#' Set the geographic location
-#' @rdname calculation
-#' @param geolon  geographic longitude as double (deg)
-#' @param geolat  geographic latitude as double (deg)
-#' @param altitude  height as double (m)
-#' @export
-swe_set_topo <- function(geolon, geolat, altitude) {
-    invisible(.Call(`_swephR_set_topo`, geolon, geolat, altitude))
+swe_fixstar2_mag <- function(starname) {
+    .Call(`_swephR_fixstar2_mag`, starname)
 }
 
 #' @title Expert interface: Eclipses and planetary phenomena
@@ -227,25 +277,39 @@ swe_refrac_extended <- function(InAlt, geoheight, atpress, attemp, lapse_rate, c
 #' @name expert-interface
 NULL
 
-#' Compute day of week
+#' Determine Julian day number from calendar date
+#' @param year  Year as interger
+#' @param month  Month as interger
+#' @param day  Day as interger
+#' @param hour  Hour as double
+#' @param gregflag  calendar type (SE_JUL_CAL=0 or SE_GREG_CAL=1)
 #' @rdname expert-interface
 #' @export
-swe_day_of_week <- function(jd) {
-    .Call(`_swephR_day_of_week`, jd)
+swe_julday <- function(year, month, day, hour, gregflag) {
+    .Call(`_swephR_julday`, year, month, day, hour, gregflag)
 }
+
+#' @title Expert interface: miscellanious functions
+#'
+#' @description For experts that are familiar with the C API of the underlying
+#'   \code{libswe}, an expert interface that mimicks that interface as closely
+#'   as possbile is provided.
+#' @param jd  Julian day number as numeric vector (day)
+#' @param t_acc tidal acceleration as double (arcsec/century^2)
+#' @param path  directory as string, where the ephemeris files are stored
+#' @param delta_t DeltaT (day)
+#' @return \code{swe_day_of_week} retruns the day of week as integer vector (0 Monday .. 6 Sunday)
+#' @return \code{swe_get_tid_acc} returns the tidal acceleration as double (arcsec/century^2)
+#' @return \code{swe_version} returns the Swiss Ephemeris version number as string
+#' @return \code{swe_deltat} returns the DeltaT as double (day)
+#' @name expert-interface
+NULL
 
 #' Get the present configured tidal acceleration
 #' @rdname expert-interface
 #' @export
 swe_get_tid_acc <- function() {
     .Call(`_swephR_get_tid_acc`)
-}
-
-#' Get the Swiss Ephemeris version number
-#' @rdname expert-interface
-#' @export
-swe_version <- function() {
-    .Call(`_swephR_version`)
 }
 
 #' Set the tidal acceleration
@@ -260,25 +324,6 @@ swe_set_tid_acc <- function(t_acc) {
 #' @export
 swe_deltat <- function(jd_ut) {
     .Call(`_swephR_deltat`, jd_ut)
-}
-
-#' Set the directory for the sefstars.txt, swe_deltat.txt and jpl files
-#' @rdname expert-interface
-#' @export
-swe_set_ephe_path <- function(path) {
-    invisible(.Call(`_swephR_set_ephe_path`, path))
-}
-
-#' Determine Julian day number from calendar date
-#' @param year  Year as interger
-#' @param month  Month as interger
-#' @param day  Day as interger
-#' @param hour  Hour as double
-#' @param gregflag  calendar type (SE_JUL_CAL=0 or SE_GREG_CAL=1)
-#' @rdname expert-interface
-#' @export
-swe_julday <- function(year, month, day, hour, gregflag) {
-    .Call(`_swephR_julday`, year, month, day, hour, gregflag)
 }
 
 #' Set one's own DeltaT
@@ -298,10 +343,13 @@ swe_deltat_ex <- function(jd_ut, ephe_flag) {
     .Call(`_swephR_deltat_ex`, jd_ut, ephe_flag)
 }
 
-#' Close Swiss Ephemeris files
-#' @rdname expert-interface
+#' Set the geographic location
+#' @rdname Section9
+#' @param geolon  geographic longitude as double (deg)
+#' @param geolat  geographic latitude as double (deg)
+#' @param altitude  height as double (m)
 #' @export
-swe_close <- function() {
-    invisible(.Call(`_swephR_close`))
+swe_set_topo <- function(geolon, geolat, altitude) {
+    invisible(.Call(`_swephR_set_topo`, geolon, geolat, altitude))
 }
