+++ man/calculation.Rd
@@ -1,17 +1,14 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/RcppExports.R, R/calculation.R
+% Please edit documentation in R/RcppExports.R, R/Section2.R, R/Section4.R
 \name{swe_fixstar2_mag}
 \alias{swe_fixstar2_mag}
-\alias{swe_set_topo}
 \alias{swe_calc}
 \alias{swe_calc_ut}
 \alias{swe_fixstar2}
 \alias{swe_fixstar2_ut}
 \title{Calculation magnitude of star}
 \usage{
-swe_fixstar2_mag(star)
-
-swe_set_topo(geolon, geolat, altitude)
+swe_fixstar2_mag(starname)
 
 swe_calc(jd_et, ipl, iflag)
 
@@ -21,16 +18,9 @@ swe_fixstar2(star, jd_et, iflag)
 
 swe_fixstar2_ut(star, jd_ut, iflag)
 }
-\arguments{
-\item{geolon}{geographic longitude as double (deg)}
-
-\item{geolat}{geographic latitude as double (deg)}
-
-\item{altitude}{height as double (m)}
-}
 \value{
 \code{swe_fixstar2_mag} returns a list with named entries \code{return} status flag as interger,
-        \code{star} updated star name as string, \code{mag} magnitude of star as double, and \code{serr} for error message as string.
+        \code{starname} updated star name as string, \code{mag} magnitude of star as double, and \code{serr} for error message as string.
 
 \code{swe_calc} returns a list with named entries \code{return} status flag as integer,
         \code{xx} updated planet name as string, and \code{serr} error message as string.
@@ -46,8 +36,6 @@ swe_fixstar2_ut(star, jd_ut, iflag)
 }
 \description{
 Calculation magnitude of star
-
-Set the geographic location
 
 Compute information of planet (ET)
